+++ CHANGELOG.md
@@ -10,3 +10,17 @@
 ## 0.0.3
 
 - 修复了在 `pd_load_state_configure.dart` 文件中的命名重复问题.
+
+## 0.0.4
+
+- 修复`PdLoadStateConfigure`类中的全局状态视图问题.
+- 在`README`文件中添加了更多使用说明.
+- 去除了`PDLoadStateLayout`类中的`padding`、`width`、`height`等属性, 交给其父组件处理.
+
+## 0.0.5
+
+- 撤回了`0.0.4`版本中的更改, 重新添加了`PDLoadStateLayout`类中的`padding`、`width`、`height`等属性.
+
+## 0.0.6
+
+- 修复了`PdLoadStateConfigure`类中的加载中视图`loadingWidgetBuilder`无法正常使用的问题.